gcp/requirements.txt
google-cloud-storage==1.42.0
tensorflow==2.11.0
Pillow
numpy
Flask==2.1.0
flask-cors==3.0.10
opencv-python==4.5.4.60
